Refactor SkippedWeeks DateTime handling
Changed the SkippedWeek property from public to private and introduced a getter and setter. This change addresses the time zone handling issues when transforming DateTime to Calendar Week. Now, it assigns the value in Universal Time and retrieves it in Local Time to ensure accurate output regardless of the server location.